requirements.txt
boto3>=1.34,<1.35
streamlit==1.34,<2.0
email-validator==2.1.0.post1
Pillow==9.0.1
python-dotenv==1.0.0
streamlit==1.28.1
streamlit-js-eval==0.1.5
